requirements.txt
Flask==3.0.0
Flask-Caching==2.1.0
humanize==4.8.0
pyinstaller==6.1.0
PyJWT==2.8.0
pymongo==4.5.0
requests==2.31.0
BeautifulSoup4
flask-cors
